service/requirements.txt
Flask
celery[redis]
pyyaml
